remove compiler warning about comparing integers of different types
Former-commit-id: cca387dad868fa1567cf2ff9cb6374006a4237cb
Former-commit-id: 6206fe2431b4372632714a4c819a47d4fa3c9d0d

+++ src/simulation/onboardDataHandling/transmitter/simpleTransmitter.cpp
@@ -127,7 +127,7 @@ int SimpleTransmitter::getMaxIndex() {
     //! - Loop through storedData, select index w/ largest amt
     double maxVal = -1.0;
     int maxIndex = -1;
-    for (int i = 0; i < this->storageUnitMsgs.back().storedData.size(); i++) {
+    for (long unsigned int i = 0; i < this->storageUnitMsgs.back().storedData.size(); i++) {
         if (this->storageUnitMsgs.back().storedData[i] > maxVal) {
             maxVal = this->storageUnitMsgs.back().storedData[i];
             maxIndex = i;
